first good version. Now we have (elementary) auth system

+++ satisfaction/praises/views.py
@@ -9,13 +9,18 @@ from .models import Praise
 
 # Create your views here.
 def index(request, date_text=timezone.localtime().strftime("%Y%m%d")):
-    latest_praise_list = Praise.objects.order_by('-pub_date')[:5]
+    if (not request.user.is_authenticated):
+        return HttpResponseRedirect(reverse("praises:login"))
+
+    # latest_praise_list = Praise.objects.order_by('-pub_date')[:5]
+
+    user_praises = Praise.objects.filter(author=request.user)
 
     date = datetime.strptime(str(date_text), "%Y%m%d").date()
-    praise_list = Praise.objects.filter(pub_date__date=date)
+    praise_list = user_praises.filter(pub_date__date=date)
 
     context = {
-        'latest_praise_list': latest_praise_list,
+        # 'latest_praise_list': latest_praise_list,
         'praise_list': praise_list,
         'date': date.strftime("%Y.%m.%d")
     }
@@ -30,7 +35,7 @@ def new(request):
     date_text = request.POST['date_text']
     date = datetime.strptime(str(date_text), "%Y.%m.%d")
 
-    p = Praise(praise_text=praise_text, pub_date=date)
+    p = Praise(praise_text=praise_text, pub_date=date, author=request.user)
     p.save()
 
     return HttpResponseRedirect(reverse("praises:index",
@@ -40,6 +45,12 @@ def new(request):
 def delete(request, praise_id):
     p = get_object_or_404(Praise, pk=praise_id)
     date = timezone.localtime(p.pub_date)
-    p.delete()
+
+    if (p.author == request.user):
+        p.delete()
+
     return HttpResponseRedirect(reverse("praises:index",
                                 kwargs={'date_text': date.strftime("%Y%m%d")}))
+
+def login(request):
+    return render(request, 'praises/login.html')
